CoSim/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "co-sim-platform"
version = "0.1.0"
description = "Control plane services for the CoSim cloud robotics platform"
authors = [{name = "CoSim"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pydantic[email]>=2.6",
    "pydantic-settings>=2.2",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.0.1,<5",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "psycopg[binary]>=3.1",
    "redis>=5.0",
    "structlog>=24.1",
    "nats-py>=2.6",
    "tenacity>=8.2",
    # Auth0 integration
    "auth0-python>=4.7",
    # Simulation dependencies
    "numpy>=1.24",
    "pillow>=10.0",
    "mujoco>=3.0",
    "pybullet>=3.2"
]

[project.optional-dependencies]
test = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "fakeredis>=2.23",
    "aiosqlite>=0.19",
    "greenlet>=3.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
